Refactor Addition/Subtraction type classes
Have arguments be type constructors, not types; improves type inference when working with types such as 'Point2d coordinates' (where the 'coordinates' type variable is unknown/arbitrary).

This required adding a 'Count units' type that acts just like 'Quantity units' but for integral values, so that it can be used with the new typeclasses.