Add ApproximateEquality instances for Direction#d
Not quite as useful as other instances since the tolerance needs to be an angle instead of a length (so can't use a single implicit tolerance and thread it through everywhere).

Might want to introduce a separate AngleTolerance type somehow, but then it would be weird that to compare two Angle values you didn't use an AngleTolerance...probably cleaner to keep things as is for now.