Remove ApproximateEquality functional dependencies
Will allow it to be implemented for types like Int or Bool,
or other types with exact equality semantics
that therefore don't actually use a tolerance with units
(useful in turn for supporting ApproximateEquality for tuples,
by implementing ApproximateEquality for all members
whether they actually need a tolerance or not).

Removing the functional dependencies
doesn't currently seem to cause any issues with type inference,
but if necessary they could probably be brought back in the future
(with some workarounds for approximate equality of tuples).