Refactor curve and function types to use Expression wherever possible
Means removing Arc and Line patterns from Curve2d for now, until those can be reimplemented in terms of checks based purely on derivative relationships etc. (e.g. a curve is a line if and only if every point on the line is equal to a linear interpolation between the curve endpoints...for quadratic splines, check first and second derivatives at curve start and then see if the whole curve is consistent with a quadratic spline with those derivatives?)