Refactor to avoid huge number of typeclass instances in Expression module
Just support Expression-Expression operations, not e.g. mixed Qty-Expression operations; makes client code more verbose in a couple places but dramatically reduces the size of the Expression module (which is getting pretty huge).

In general, treat the Expression type/module as an "internal implementation detail" that most actual client code won't use, so it's OK for it to not have quite the same 'niceties' as other types/modules.