Make type definitions more consistent
Use 'space @ units' instead of 'coordinateSystem' everywhere, for consistency - seems to work out the best for nice type inference etc. Possibly a performance regression due to more expensive units coercion, but if necessary it should be possible to explicitly implement Multiplication and Division typeclasses (instead of relying on the default implementation which uses units coercion).