Remove unnecessary Qty.abs calls
Second derivative *must* be positive in these cases (otherwise the function in the neighbourhood of the tangent point being tested would be negative, not positive).
And Qty.sqrt already clamps its input to zero, so small amounts of roundoff causing the second derivative value to be slightly negative won't cause issues.