Re-add World3d module
Using Frame3d.world everywhere is a bit annoying and adds an extra level of indirection to the code;
also to make it ergonomic requires having e.g. frontPlane being a *field* on Frame3d,
which in turns requires frontPlane to have a non-polymorphic return type
(so that the field type can be inferred from the parent type, as required by HasField)
which in turn requires the weird 'FrontPlane space' etc. predefined generic spaces