Add Tolerance.forEndpointDerivative
Figure out reasonable scaling factor for tolerance endpoint *derivatives*,
since derivative values tend to blow up to pretty large values for higher and higher order derivatives.
For example, for the tolerance on the first derivative,
imagine a quadratic that is zero at t=0 and t=1,
i.e. f(t) = k * t * (t - 1) for some constant k.
(This is, roughly, the simplest polynomial that is zero at t=0 and t=1
but has non-zero derivatives at those points.)
Then, solve for k such that the maximum value of that function (the value at t=0.5)
is equal to a desired tolearnce (epsilon) value.
Finally, evaluate the first derivative at t=0 (or t=1, they will be the same magnitude);
it works out that the derivative will be 4 times the desired tolerance.
Therefore, we can conclude that with a reasonable approximation to the curve between the two (assumed zero) endpoints,
the curve will always be zero within the desired tolerance
if the first derivatives at both endpoints are zero to within 4 times the desired tolerance.

The logic can be extende to arbitrary derivative orders, e.g. f(t) = k * t^n * (t - 1)^n,
with the scaling factor turning tou to be equal to n! * 2^(2 * n) in the general case.