Remove Vertex2d class
Just use Point2d; simpler in general and much easier to map to target languages.

Required simplifying Body3d meshing logic,
which means that surface vertices along edges are no longer guaranteed to *exactly* match the vertex on the neighbouring surface,
but any difference should be within numerical roundoff
(so not visible when rendering, and easy to merge as a post-processing step if exact watertighness is desired)
and the simpler logic should be more efficient.